+++ src/applications/Filippovich/db/bestPlayers.json
@@ -68,6 +68,13 @@
       "difficulty": "Easy",
       "score": 165,
       "id": 10
+    },
+    {
+      "name": "fbg",
+      "steps": 457,
+      "difficulty": "Medium",
+      "score": 914,
+      "id": 11
     }
   ]
 }
