+++ src/applications/Filippovich/src/containers/gameLog.js
@@ -1,0 +1,10 @@
+import {connect} from 'react-redux';
+import GameLog from '../components/gameLog';
+
+const mapStateToProps = state => ({
+    logText: state.games.logText,
+    movesCount: state.moves.movesCount,
+});
+
+
+export default connect(mapStateToProps)(GameLog);
