+++ src/applications/Filippovich/src/containers/stats.js
@@ -1,0 +1,20 @@
+import {connect} from "react-redux";
+import Stats from '../components/stats';
+
+
+const mapStateToProps = state => ({
+    name: state.settings.inputNameValue,
+    level: state.games.level,
+    health: state.games.health,
+    experience: state.games.experience,
+    experienceLeftToCollect: state.games.experienceLeftToCollect,
+    skills: state.games.skills,
+    skillsLeftToCollect: state.games.skillsLeftToCollect,
+    certifications: state.games.certifications,
+    certificationsLeftToCollect: state.games.certificationsLeftToCollect,
+    ultimate: state.games.ultimate,
+    ultimateLeftToCollect: state.games.ultimateLeftToCollect,
+});
+
+
+export default connect(mapStateToProps)(Stats);
